added saving large matrix to binary file option in Extract Value;fixed a bug of loading/saving hdr/img files

+++ brant_pub/MODIFIED_NIFTI_20160621/load_nii_ext_mod.m
@@ -183,7 +183,7 @@ if vox_offset
     end_of_ext = vox_offset;
 else % no voxels, all extension
     if foptgz(fileprefix, 'remain', uint32(348)) > 0
-        end_of_ext = 352 + foptgz(fileprefix, 'remain', uint32(352));
+        end_of_ext = 352 + double(foptgz(fileprefix, 'remain', uint32(352)));
     else
         end_of_ext = 348;
     end
